backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
sqlmodel==0.0.22
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
google-generativeai==0.7.2
opencv-python
mediapipe
passlib[bcrypt]
python-jose[cryptography]
Pillow
numpy
bcrypt==4.3.0
